Added method for single employee

+++ src/main/java/com/poc/springboot/employeedetails/EmployeeSource.java
@@ -31,7 +31,7 @@ public class EmployeeSource {
         return EMPLOYEES.values();
     }
     
-    public static Employee getEmployee(long empId) {
+    public static Employee getEmployee(Integer empId) {
         return EMPLOYEES.get(empId);
     }
 }
